Implement further rolling calculation
JS - add an "or" conditional to handleOperator function to work in conjunction with calculate function. If two operands are added followed by pressing the "=" sign to fire the calculate function, there will be a check through handleOperator to see if there is only an assignment to previousNum, in which that variable assignment will be passed on to calculateRolling function. This ensures that a user can continue to perform additional math on the result after the calculate function is fired with the "=" button